HW-4 MovieItem 의 String 에 html 태그 없애는 함수 Adapter 로 이동
