+++ 2004/nukeolaf/app/src/main/java/com/olaf/nukeolaf/data/model/MovieItem.kt
@@ -5,12 +5,12 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class MovieItem(
-    val title: String,
+    var title: String,
     val link: String,
     val image: String,
-    val subtitle: String,
+    var subtitle: String,
     val pubDate: String,
-    val director: String,
-    val actor: String,
+    var director: String,
+    var actor: String,
     val userRating: Float
 ) : Parcelable
